Fixed manifest merge issue (#19)
The application element in the manifest is not needed and forces users of this library to utilize tools:replace="android:allowBackup" when switching to android:allowBackup="false" in their manifest. allowBackup is true by default in any case. Focusing RTL with supportsRtl="true" might be less likely to cause an issue but is just unnecessary.